changed structure of full spot list
